requirements.txt
python-dotenv
pandas
SQLAlchemy